+++ src/ballistica/shared/ballistica.cc
@@ -39,8 +39,8 @@ auto main(int argc, char** argv) -> int {
 namespace ballistica {
 
 // These are set automatically via script; don't modify them here.
-const int kEngineBuildNumber = 22318;
-const char* kEngineVersion = "1.7.38";
+const int kEngineBuildNumber = 22320;
+const char* kEngineVersion = "1.7.39";
 const int kEngineApiVersion = 9;
 
 #if BA_MONOLITHIC_BUILD