+++ tools/efro/dataclassio/_base.py
@@ -155,42 +155,43 @@ class IOMultiType(Generic[EnumT]):
 class IOAttrs:
     """For specifying io behavior in annotations.
 
-    'storagename', if passed, is the name used when storing to json/etc.
+    :param storagename: If passed, is the name used when storing to
+        json/etc.
 
-    'store_default' can be set to False to avoid writing values when
-        equal to the default value. Note that this requires the
+    :param store_default: Can be set to False to avoid writing values
+        when equal to the default value. Note that this requires the
         dataclass field to define a default or default_factory or for
         its IOAttrs to define a soft_default value.
 
-    'whole_days', if True, requires datetime values to be exactly on day
+    :param whole_days: If True, requires datetime values to be exactly on day
         boundaries (see efro.util.utc_today()).
 
-    'whole_hours', if True, requires datetime values to lie exactly on
-        hour boundaries (see efro.util.utc_this_hour()).
+    :param whole_hours: If True, requires datetime values to lie exactly
+        on hour boundaries (see efro.util.utc_this_hour()).
 
-    'whole_minutes', if True, requires datetime values to lie exactly on
-        minute boundaries (see efro.util.utc_this_minute()).
+    :param whole_minutes: If True, requires datetime values to lie
+        exactly on minute boundaries (see efro.util.utc_this_minute()).
 
-    'soft_default', if passed, injects a default value into dataclass
-        instantiation when the field is not present in the input data.
-        This allows dataclasses to add new non-optional fields while
-        gracefully 'upgrading' old data. Note that when a soft_default
-        is present it will take precedence over field defaults when
-        determining whether to store a value for a field with
-        store_default=False (since the soft_default value is what we'll
-        get when reading that same data back in when the field is
+    :param soft_default: If passed, injects a default value into
+        dataclass instantiation when the field is not present in the
+        input data. This allows dataclasses to add new non-optional
+        fields while gracefully 'upgrading' old data. Note that when a
+        soft_default is present it will take precedence over field
+        defaults when determining whether to store a value for a field
+        with store_default=False (since the soft_default value is what
+        we'll get when reading that same data back in when the field is
         omitted).
 
-    'soft_default_factory' is similar to 'default_factory' in dataclass
-        fields; it should be used instead of 'soft_default' for mutable
-        types such as lists to prevent a single default object from
-        unintentionally changing over time.
+    :param soft_default_factory: Is similar to 'default_factory' in
+        dataclass fields; it should be used instead of 'soft_default'
+        for mutable types such as lists to prevent a single default
+        object from unintentionally changing over time.
 
-    'enum_fallback', if provided, specifies an enum value that can be
-        substituted in the case of unrecognized input values. This can
-        allow newer data to remain loadable in older environments. Note
-        that 'lossy' must be enabled in the top level load call for this
-        to apply, since it can fundamentally modify data.
+    :param enum_fallback: If provided, specifies an enum value that can
+        be substituted in the case of unrecognized input values. This
+        can allow newer data to remain loadable in older environments.
+        Note that 'lossy' must be enabled in the top level load call for
+        this to apply, since it can fundamentally modify data.
     """
 
     # A sentinel object to detect if a parameter is supplied or not. Use